LoadCell ino file with wifi instead of ethernet
Made some changes where the mqtt doesn't reconnect constantly. Only when active connection. For my sitiuation i measure the bed. Since if you move around at night it will do a lot of updates so i changed the frequency and only full kilo's. This results in less traffic and fewer updates. (it detects that i am in bed in a few seconds, thats fast enough)

Also copy pasted the over the air update from the other repo in your readme. Reason to adjust this one was that the other one crashed every now and then. I redid the wifi part and it seems to be solid (for now)

Last part is that if the value is a negative value i just reset it back to 0. This removes weird spikes if nobody is in the bed and saves false calls too. 

Hope it helps someone ;)